+++ lib/smart_bottom_sheet.dart
@@ -19,3 +19,6 @@ export 'src/sheets/confirm_sheet.dart';
 export 'src/sheets/stepper_sheet.dart';
 export 'src/sheets/rating_sheet.dart';
 export 'src/utils/sheet_config.dart';
+export 'src/utils/sheet_stack_manager.dart';
+export 'src/utils/sheet_handle.dart';
+